+++ projects/net.splitcells.gel.sheath/src/main/java/net/splitcells/gel/test/functionality/NQueenProblemTest.java
@@ -78,7 +78,7 @@ public class NQueenProblemTest extends TestSuiteI {
         assertThat(testSubject.constraint().rating()).isEqualTo(cost(0));
     }
 
-    @Disabled
+    //@Disabled
     @Tag(CAPABILITY_TEST)
     @Test
     public void test_8_queen_problem_with_backtracking() {